Fix NextAuth route handler configuration
